backend/requirements.txt
yfinance
pandas-ta
pymongo
groq
google-genai
python-dotenv
pandas
requests
flask
flask-cors
